add status for task

+++ src/main/java/com/quangduong/SE114backend/service/impl/TaskServiceImpl.java
@@ -4,19 +4,23 @@ import com.quangduong.SE114backend.dto.task.TaskDTO;
 import com.quangduong.SE114backend.dto.task.TaskDetailsDTO;
 import com.quangduong.SE114backend.dto.task.TaskUpdateDTO;
 import com.quangduong.SE114backend.entity.TaskEntity;
+import com.quangduong.SE114backend.exception.NoPermissionException;
 import com.quangduong.SE114backend.exception.ResourceNotFoundException;
 import com.quangduong.SE114backend.mapper.TaskMapper;
 import com.quangduong.SE114backend.repository.sql.*;
 import com.quangduong.SE114backend.service.TaskService;
+import com.quangduong.SE114backend.utils.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class TaskServiceImpl implements TaskService {
+
+    @Autowired
+    private SecurityUtils securityUtils;
 
     @Autowired
     private TaskRepository taskRepository;
@@ -52,26 +56,26 @@ public class TaskServiceImpl implements TaskService {
     public TaskDetailsDTO updateTask(TaskUpdateDTO dto) {
         TaskEntity entity = taskRepository.findById(dto.getId())
                 .orElseThrow(() -> new ResourceNotFoundException("Not found task with id: " + dto.getId()));
-        List<Long> textAttributeIds = entity.getTextAttributes().stream().map(t -> t.getId()).collect(Collectors.toList());
-        List<Long> numberAttributeIds = entity.getNumberAttributes().stream().map(t -> t.getId()).collect(Collectors.toList());
-        List<Long> dateAttributeIds = entity.getDateAttributes().stream().map(t -> t.getId()).collect(Collectors.toList());
-        List<Long> labelAttributeIds = entity.getLabelAttributes().stream().map(t -> t.getId()).collect(Collectors.toList());
+        List<Long> textAttributeIds = entity.getTextAttributes().stream().map(t -> t.getId()).toList();
+        List<Long> numberAttributeIds = entity.getNumberAttributes().stream().map(t -> t.getId()).toList();
+        List<Long> dateAttributeIds = entity.getDateAttributes().stream().map(t -> t.getId()).toList();
+        List<Long> labelAttributeIds = entity.getLabelAttributes().stream().map(t -> t.getId()).toList();
         entity = taskMapper.toEntity(dto, entity);
         TaskEntity finalEntity = entity;
         textAttributeIds.forEach(i -> {
-            if(finalEntity.getTextAttributes().stream().noneMatch(t -> t.getId() == i))
+            if (finalEntity.getTextAttributes().stream().noneMatch(t -> t.getId() == i))
                 textAttributeRepository.deleteById(i);
         });
         numberAttributeIds.forEach(i -> {
-            if(finalEntity.getNumberAttributes().stream().noneMatch(t -> t.getId() == i))
+            if (finalEntity.getNumberAttributes().stream().noneMatch(t -> t.getId() == i))
                 numberAttributeRepository.deleteById(i);
         });
         dateAttributeIds.forEach(i -> {
-            if(finalEntity.getDateAttributes().stream().noneMatch(t -> t.getId() == i))
+            if (finalEntity.getDateAttributes().stream().noneMatch(t -> t.getId() == i))
                 dateAttributeRepository.deleteById(i);
         });
         labelAttributeIds.forEach(i -> {
-            if(finalEntity.getLabelAttributes().stream().noneMatch(t -> t.getId() == i))
+            if (finalEntity.getLabelAttributes().stream().noneMatch(t -> t.getId() == i))
                 labelAttributeRepository.deleteById(i);
         });
         taskRepository.save(entity);
@@ -81,6 +85,11 @@ public class TaskServiceImpl implements TaskService {
     @Override
     @Transactional
     public void deleteTask(long id) {
+        TaskEntity entity = taskRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Not found task with id: " + id));
+        if (entity.getTable().getMembers().stream().noneMatch(m -> m.getId() == securityUtils.getCurrentUserId())
+                && !entity.getTable().getCreatedBy().equals(securityUtils.getCurrentUser().getEmail()))
+            throw new NoPermissionException("Not allowed");
         taskRepository.deleteById(id);
     }
 }
